Lessons/Lesson 18/dspy-moby/requirements.txt
dspy-ai
gradio
faiss-cpu
google-generativeai
sentence-transformers
python-dotenv
